Refactor resource module into cohesive packages
