+++ falcon_pachinko/dispatcher.py
@@ -1,0 +1,127 @@
+"""Message dispatch helpers for :class:`WebSocketResource`."""
+
+from __future__ import annotations
+
+import inspect
+import typing
+
+import msgspec
+import msgspec.inspect as msgspec_inspect
+import msgspec.json as msjson
+
+from .exceptions import (
+    HandlerNotAsyncError,
+    HandlerSignatureError,
+    SignatureInspectionError,
+)
+from .handlers import Handler, HandlerInfo, get_payload_type
+from .schema import requires_strict_validation, validate_strict_payload
+from .utils import to_snake_case
+
+if typing.TYPE_CHECKING:  # pragma: no cover - type hints only
+    from .protocols import WebSocketLike
+    from .resource import WebSocketResource
+
+
+class Envelope(msgspec.Struct, frozen=True):
+    """Simple envelope for messages without a schema."""
+
+    type: str
+    payload: typing.Any | None = None
+
+
+def find_conventional_handler(
+    resource: WebSocketResource, tag: str
+) -> HandlerInfo | None:
+    """Return a handler matching ``on_{tag}`` if present."""
+    name = f"on_{to_snake_case(tag)}"
+    func = getattr(resource.__class__, name, None)
+    if func is None or not inspect.iscoroutinefunction(func):
+        return None
+    try:
+        payload_type = get_payload_type(typing.cast("Handler", func))
+    except (HandlerSignatureError, HandlerNotAsyncError, SignatureInspectionError):
+        return None
+    return HandlerInfo(typing.cast("Handler", func), payload_type, strict=True)
+
+
+async def convert_and_invoke_handler(
+    resource: WebSocketResource,
+    ws: WebSocketLike,
+    raw: str | bytes,
+    handler_info: HandlerInfo,
+    payload: object,
+) -> None:
+    """Convert ``payload`` to the handler's type and invoke it."""
+    payload_type = handler_info.payload_type
+    if payload_type is not None and payload is not None:
+        try:
+            if requires_strict_validation(
+                payload, payload_type, strict=handler_info.strict
+            ):
+                validate_strict_payload(
+                    payload, payload_type, strict=handler_info.strict
+                )
+            payload = typing.cast(
+                "typing.Any",
+                msgspec.convert(payload, type=payload_type, strict=handler_info.strict),
+            )
+        except msgspec.ValidationError:
+            await resource.on_unhandled(ws, raw)
+            return
+    await handler_info.handler(resource, ws, payload)
+
+
+async def dispatch(
+    resource: WebSocketResource, ws: WebSocketLike, raw: str | bytes
+) -> None:
+    """Dispatch ``raw`` to a handler based on :attr:`resource.schema`."""
+    if resource.schema is not None:
+        await dispatch_with_schema(resource, ws, raw)
+    else:
+        await dispatch_with_envelope(resource, ws, raw)
+
+
+async def dispatch_with_schema(
+    resource: WebSocketResource, ws: WebSocketLike, raw: str | bytes
+) -> None:
+    """Decode and dispatch ``raw`` using ``resource.schema``."""
+    try:
+        message = msjson.decode(raw, type=resource.schema)
+    except (msgspec.DecodeError, msgspec.ValidationError):
+        await resource.on_unhandled(ws, raw)
+        return
+
+    entry = resource.__class__._struct_handlers.get(type(message))
+    if not entry:
+        info = msgspec_inspect.type_info(type(message))
+        tag_val = typing.cast("msgspec_inspect.StructType", info).tag
+        conv = find_conventional_handler(resource, typing.cast("str", tag_val))
+        if conv is None:
+            await resource.on_unhandled(ws, raw)
+            return
+        await convert_and_invoke_handler(resource, ws, raw, conv, message)
+        return
+
+    await convert_and_invoke_handler(resource, ws, raw, entry, message)
+
+
+async def dispatch_with_envelope(
+    resource: WebSocketResource, ws: WebSocketLike, raw: str | bytes
+) -> None:
+    """Decode and dispatch ``raw`` using the envelope format."""
+    try:
+        envelope = msjson.decode(raw, type=Envelope)
+    except msgspec.DecodeError:
+        await resource.on_unhandled(ws, raw)
+        return
+
+    handler_entry = resource.__class__.handlers.get(envelope.type)
+    if handler_entry is None:
+        handler_entry = find_conventional_handler(resource, envelope.type)
+
+    if handler_entry is None:
+        await resource.on_unhandled(ws, raw)
+        return
+
+    await convert_and_invoke_handler(resource, ws, raw, handler_entry, envelope.payload)
